Upgrad USI to 0.1.9.
Summary:
This enables us to define custom container infos for Mesos tasks.
