+++ src/main/java/org/jenkinsci/plugins/mesos/MesosApi.java
@@ -14,6 +14,9 @@ import com.mesosphere.mesos.conf.MesosClientSettings;
 import com.mesosphere.usi.core.conf.SchedulerSettings;
 import com.mesosphere.usi.core.japi.Scheduler;
 import com.mesosphere.usi.core.models.*;
+import com.mesosphere.usi.core.models.commands.KillPod;
+import com.mesosphere.usi.core.models.commands.LaunchPod;
+import com.mesosphere.usi.core.models.commands.SchedulerCommand;
 import com.mesosphere.usi.repository.PodRecordRepository;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
