+++ src/test/java/org/jenkinsci/plugins/mesos/integration/MesosApiTest.java
@@ -10,8 +10,8 @@ import akka.NotUsed;
 import akka.actor.ActorSystem;
 import akka.stream.ActorMaterializer;
 import akka.stream.javadsl.Flow;
-import com.mesosphere.usi.core.models.SchedulerCommand;
 import com.mesosphere.usi.core.models.StateEvent;
+import com.mesosphere.usi.core.models.commands.SchedulerCommand;
 import com.mesosphere.utils.mesos.MesosClusterExtension;
 import com.mesosphere.utils.zookeeper.ZookeeperServerExtension;
 import hudson.model.Descriptor.FormException;
